requirements.txt
Cython==0.23.4
Jinja2==2.8
Keras==0.3.2
MarkupSafe==0.23
Pillow==3.1.1
PyYAML==3.11
Pygments==2.1.3
Theano==0.8.0rc1
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
catkin==0.6.11
certifi==2016.2.28
controller-manager-msgs==0.9.1
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
diagnostic-updater==1.8.7
functools32==3.2.3.post2
gencpp==0.5.3
genlisp==0.4.15
genmsg==0.5.6
genpy==0.5.4
h5py==2.5.0
image-geometry==1.11.7
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
matplotlib==1.5.1
message-filters==1.11.10
mistune==0.7.2
mpmath==0.19
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
notebook==4.1.0
numexpr==2.5
numpy==1.10.4
pandas==0.18.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
pylab==0.1.3
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2016.1
pyzmq==15.2.0
qtconsole==4.2.0
rosboost-cfg==1.11.6
rosclean==1.11.6
roscreate==1.11.6
rosgraph==1.11.10
roslint==0.9.3
roslz4==1.11.10
rosmake==1.11.6
rosmaster==1.11.10
rosparam==1.11.10
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
seaborn==0.7.0
sensor-msgs==1.11.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
smach==2.0.0
smclib==1.7.16
sympy==1.0
tables==3.2.2
terminado==0.6
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
wsgiref==0.1.2
